EES-5446 - grouped long-running test functions into single class
